+++ src/model/song.model.ts
@@ -31,3 +31,9 @@ export interface UserSong extends Song {
     fav?: boolean;
     tags?: string[];
 }
+
+export interface UserSongFilter {
+    limit: number;
+    status: Record<Status, boolean>;
+    fav: boolean | undefined;
+}